Update ext-token.json
Hi, I have added a categorisation to the tokens! I think it could help new users?
It not always easy to split between the function of the project and the function of the token (HeN is NFT but hdao will be used for governance for example)
We can discuss it if you find it interesting?